refactor(ably-subscriber): split lib.rs into separate modules
